test(adapters): cover remote tool retry policy wrappers
