+++ pyagentspec/tests/adapters/langgraph/test_tools.py
@@ -4,7 +4,6 @@
 # (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0) or Universal Permissive License
 # (UPL) 1.0 (LICENSE-UPL or https://oss.oracle.com/licenses/upl), at your option.
 
-import ssl
 import threading
 from typing import Any
 from unittest.mock import patch
@@ -21,7 +20,6 @@ from pyagentspec.flows.nodes.endnode import EndNode
 from pyagentspec.flows.nodes.startnode import StartNode
 from pyagentspec.llms import OpenAiCompatibleConfig
 from pyagentspec.property import IntegerProperty, Property
-from pyagentspec.retrypolicy import RetryPolicy
 from pyagentspec.tools import ClientTool, RemoteTool, ServerTool
 
 
@@ -319,230 +317,6 @@ def test_remote_tool_rejects_rendered_url_outside_allow_list_with_langgraph(
             lang_tool.func(host="blocked.example.com")
 
     mocked_request.assert_not_called()
-
-
-def _make_remote_tool_for_retry_tests(retry_policy: RetryPolicy | None) -> RemoteTool:
-    return RemoteTool(
-        name="retry_service",
-        description="A remote service with retry policy",
-        url="https://example.com/api",
-        http_method="GET",
-        retry_policy=retry_policy,
-    )
-
-
-@pytest.mark.parametrize(
-    "retry_policy, expected_timeout",
-    [
-        pytest.param(
-            RetryPolicy(max_attempts=0, request_timeout=300.0),
-            300.0,
-            id="retry-policy-timeout",
-        ),
-        pytest.param(
-            RetryPolicy(max_attempts=0),
-            None,
-            id="retry-policy-without-timeout",
-        ),
-        pytest.param(None, None, id="no-retry-policy"),
-    ],
-)
-def test_remote_tool_retry_policy_timeout_configuration(
-    retry_policy: RetryPolicy | None, expected_timeout: float | None
-) -> None:
-    """Verify RemoteTool retry policy timeout configuration is passed only when set."""
-    from pyagentspec.adapters.langgraph import AgentSpecLoader
-
-    remote_tool = _make_remote_tool_for_retry_tests(retry_policy)
-    lang_tool = AgentSpecLoader().load_component(remote_tool)
-
-    with patch("pyagentspec.adapters._tools_common.httpx.request") as mock_request:
-        mock_request.return_value = DummyResponse({"result": "ok"})
-        lang_tool.func()
-        _, called_kwargs = mock_request.call_args
-
-    if expected_timeout is None:
-        assert "timeout" not in called_kwargs
-    else:
-        assert isinstance(called_kwargs["timeout"], httpx.Timeout)
-        assert called_kwargs["timeout"].read == expected_timeout
-
-
-def _get_retry_success_side_effects(case_name: str) -> list[Any]:
-    request = httpx.Request("GET", "https://example.com/api")
-    if case_name == "transport-error":
-        return [
-            httpx.ConnectError("temporary failure"),
-            httpx.ConnectError("temporary failure"),
-            DummyResponse({"result": "ok"}),
-        ]
-    if case_name == "service-error":
-        return [
-            httpx.Response(503, request=request, json={"error": "busy"}),
-            DummyResponse({"result": "ok"}),
-        ]
-    if case_name == "status-text-code":
-        return [
-            httpx.Response(
-                429,
-                request=request,
-                json={"code": "TooManyRequests", "message": "throttled"},
-            ),
-            DummyResponse({"result": "ok"}),
-        ]
-    raise ValueError(f"Unknown retry success case: {case_name}")
-
-
-@pytest.mark.parametrize(
-    "case_name, retry_policy, expected_call_count, expected_timeout",
-    [
-        pytest.param(
-            "transport-error",
-            RetryPolicy(
-                max_attempts=2,
-                request_timeout=0.5,
-                initial_retry_delay=0,
-                max_retry_delay=0,
-            ),
-            3,
-            0.5,
-            id="transport-error",
-        ),
-        pytest.param(
-            "service-error",
-            RetryPolicy(
-                max_attempts=1,
-                initial_retry_delay=0,
-                max_retry_delay=0,
-            ),
-            2,
-            None,
-            id="default-5xx-service-error",
-        ),
-        pytest.param(
-            "status-text-code",
-            RetryPolicy(
-                max_attempts=1,
-                initial_retry_delay=0,
-                max_retry_delay=0,
-                service_error_retry_on_any_5xx=False,
-                recoverable_statuses={"429": ["TooManyRequests"]},
-            ),
-            2,
-            None,
-            id="configured-status-text-code",
-        ),
-    ],
-)
-def test_remote_tool_retry_policy_retries_then_succeeds(
-    case_name: str,
-    retry_policy: RetryPolicy,
-    expected_call_count: int,
-    expected_timeout: float | None,
-) -> None:
-    """Verify retry policy succeeds after transient transport or response failures."""
-    from pyagentspec.adapters.langgraph import AgentSpecLoader
-
-    remote_tool = _make_remote_tool_for_retry_tests(retry_policy)
-    lang_tool = AgentSpecLoader().load_component(remote_tool)
-
-    with patch("pyagentspec.adapters._tools_common.httpx.request") as mock_request:
-        mock_request.side_effect = _get_retry_success_side_effects(case_name)
-        assert lang_tool.func() == {"result": "ok"}
-        assert mock_request.call_count == expected_call_count
-
-    if expected_timeout is not None:
-        assert all(
-            isinstance(call.kwargs["timeout"], httpx.Timeout)
-            and call.kwargs["timeout"].read == expected_timeout
-            for call in mock_request.call_args_list
-        )
-    else:
-        assert all("timeout" not in call.kwargs for call in mock_request.call_args_list)
-
-
-def test_remote_tool_retry_policy_honors_retry_after_header() -> None:
-    """Verify RemoteTool retry policy prefers bounded Retry-After delays."""
-    from pyagentspec.adapters.langgraph import AgentSpecLoader
-
-    remote_tool = _make_remote_tool_for_retry_tests(
-        RetryPolicy(
-            max_attempts=1,
-            initial_retry_delay=0,
-            max_retry_delay=0,
-            recoverable_statuses={"429": []},
-        )
-    )
-
-    lang_tool = AgentSpecLoader().load_component(remote_tool)
-    request = httpx.Request("GET", "https://example.com/api")
-
-    with (
-        patch("pyagentspec.adapters._tools_common.httpx.request") as mock_request,
-        patch("pyagentspec.adapters._tools_common.time.sleep") as mock_sleep,
-    ):
-        mock_request.side_effect = [
-            httpx.Response(
-                429,
-                request=request,
-                headers={"retry-after": "45"},
-                json={"error": "throttled"},
-            ),
-            DummyResponse({"result": "ok"}),
-        ]
-        assert lang_tool.func() == {"result": "ok"}
-        mock_sleep.assert_called_once_with(30.0)
-
-
-def test_remote_tool_retry_policy_raises_after_recoverable_status_retries() -> None:
-    """Verify RemoteTool retry policy raises after recoverable responses are exhausted."""
-    from pyagentspec.adapters.langgraph import AgentSpecLoader
-
-    remote_tool = _make_remote_tool_for_retry_tests(
-        RetryPolicy(
-            max_attempts=1,
-            initial_retry_delay=0,
-            max_retry_delay=0,
-        )
-    )
-
-    lang_tool = AgentSpecLoader().load_component(remote_tool)
-    request = httpx.Request("GET", "https://example.com/api")
-
-    with patch("pyagentspec.adapters._tools_common.httpx.request") as mock_request:
-        mock_request.side_effect = [
-            httpx.Response(503, request=request, json={"error": "busy"}),
-            httpx.Response(503, request=request, json={"error": "still busy"}),
-        ]
-        with pytest.raises(httpx.HTTPStatusError, match="503"):
-            lang_tool.func()
-        assert mock_request.call_count == 2
-
-
-def test_remote_tool_retry_policy_does_not_retry_tls_errors() -> None:
-    """Verify RemoteTool retry policy does not retry TLS verification failures."""
-    from pyagentspec.adapters.langgraph import AgentSpecLoader
-
-    remote_tool = _make_remote_tool_for_retry_tests(
-        RetryPolicy(
-            max_attempts=2,
-            initial_retry_delay=0,
-            max_retry_delay=0,
-        )
-    )
-
-    lang_tool = AgentSpecLoader().load_component(remote_tool)
-    try:
-        raise ssl.SSLCertVerificationError("certificate verify failed")
-    except ssl.SSLCertVerificationError as cert_error:
-        try:
-            raise httpx.ConnectError("TLS handshake failed") from cert_error
-        except httpx.ConnectError as tls_error:
-            with patch("pyagentspec.adapters._tools_common.httpx.request") as mock_request:
-                mock_request.side_effect = tls_error
-                with pytest.raises(httpx.ConnectError, match="TLS handshake failed"):
-                    lang_tool.func()
-                assert mock_request.call_count == 1
 
 
 def _make_simple_flow_with_tool(tool_node, start_inputs=None, end_outputs=None):
